updated archive download logic

+++ tech-ops/scripts/env_precheck.sh
@@ -7,11 +7,15 @@ if [[ ${OFFLINE_MODE,,} == "true" ]]; then
         exit 1
     else
         ctx logger info "Downloading offline binary..."
-        ctx logger info "curl -sfLu "$OFFLINE_BINARY_USER:$OFFLINE_BINARY_PASSWORD" $OFFLINE_BINARY_URL -o ~/$(basename $OFFLINE_BINARY_URL)"
-        curl -skfLu "${OFFLINE_BINARY_USER}:${OFFLINE_BINARY_PASSWORD}" "${OFFLINE_BINARY_URL}" -o ~/$(basename $OFFLINE_BINARY_URL)
-        if [[ $? -ne 0 ]]; then
-            ctx logger info "Failed to download offline binary."
-            exit 1
+        ctx logger info "curl -skfLu "$OFFLINE_BINARY_USER:$OFFLINE_BINARY_PASSWORD" $OFFLINE_BINARY_URL -o ~/$(basename $OFFLINE_BINARY_URL)"
+        if [[ -f ~/$(basename $OFFLINE_BINARY_URL) ]]; then
+            ctx logger info "Offline binary already exists. Skipping download."
+        else
+            curl -skfLu "${OFFLINE_BINARY_USER}:${OFFLINE_BINARY_PASSWORD}" "${OFFLINE_BINARY_URL}" -o ~/$(basename $OFFLINE_BINARY_URL)
+            if [[ $? -ne 0 ]]; then
+                ctx logger info "Failed to download offline binary."
+                exit 1
+            fi
         fi
         ctx logger info "Extracting offline binary..."
         tar -xzf ~/$(basename $OFFLINE_BINARY_URL) -C ~/
